perf: check if inode in VFS has i_write_hint field
This commit addresses the review comments from
commit 046a3af5851b.
It includes minor fixes such as correcting field types,
refining condition checks, and adding initialization assignments.
Additionally, it enhances the PR description with
further details on FDP technology and the benchmark results
demonstrating WAF reduction.

+++ common/source/common/net/message/session/rw/WriteLocalFileMsg.h
@@ -22,7 +22,7 @@ class WriteLocalFileMsgBase
        */
       WriteLocalFileMsgBase(const NumNodeID clientNumID, const char* fileHandleID,
          const uint16_t targetID, const PathInfo* pathInfo, const unsigned accessFlags,
-         const int64_t offset, const int64_t count, const unsigned writeHint)
+         const int64_t offset, const int64_t count, const uint64_t writeHint)
       {
          this->clientNumID = clientNumID;
